+++ functions/Get-PSScriptTools.ps1
@@ -4,7 +4,10 @@
 Function Get-PSScriptTools {
     [cmdletbinding()]
     [OutputType("PSScriptTool")]
-    Param()
+    Param(
+        [Parameter(HelpMessage = "Filter commands based on a standard verb.")]
+        [string]$Verb
+    )
 
     Write-Verbose "Starting $($myinvocation.mycommand)"
 
@@ -21,16 +24,23 @@ Function Get-PSScriptTools {
     #getting commands directly from the module because for some unknown reason,
     #probably scope related, when using Get-Command alone to list commands in the module,
     #it includes private functions
+    Write-Verbose "Getting the PSScriptTools module"
     $mod = Get-Module -Name PSScriptTools
 
+    Write-Verbose "Using version $($mod.version)"
+    Write-Verbose "Getting exported functions"
     $funs = $mod.ExportedFunctions.keys | Get-Command
 
+    if ($verb) {
+        Write-Verbose "Filtering on verb $Verb"
+        $funs = $funs.where({$_.verb -eq $verb})
+    }
     Write-Verbose "Found $($funs.count) functions"
 
     foreach ($fun in $funs) {
         Write-Verbose "Processing $($fun.name)"
         #get aliases, ignoring errors for those commands without one
-        $alias = (Get-Alias -Definition $fun.Name -ErrorAction silentlycontinue).name
+        $alias = (Get-Alias -Definition $fun.Name -ErrorAction silentlycontinue).nameget
 
         [pscustomobject]@{
             PSTypeName = "PSScriptTool"
@@ -43,4 +53,16 @@ Function Get-PSScriptTools {
     }
 
     Write-Verbose "Ending $($myinvocation.mycommand)"
+}
+
+#define an argumentcompleter for the Verb parameter
+Register-ArgumentCompleter -CommandName Get-PSScriptTools -ParameterName Verb -ScriptBlock {
+    param($commandName, $parameterName, $wordToComplete, $commandAst, $fakeBoundParameter)
+
+    #PowerShell code to populate $wordtoComplete
+    Get-Verb | Where-object {$_.verb -like "$wordtocomplete*"} |
+        ForEach-Object {
+            # completion text,listitem text,result type,Tooltip
+            [System.Management.Automation.CompletionResult]::new($_.verb, $_.verb, 'ParameterValue', $_.group)
+        }
 }